+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # Change Log
 
-All notable changes to the "claude-dev" extension will be documented in this file.
 
-<!-- Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file. -->
+## [2.0.0]
+
+- New name! Meet Cline, an AI assistant that can use your CLI and Editor.
+- Responses are now streamed with a yellow text decoration animation to keep track of Cline's progress as he edits files.
+- New Cancel button to give Cline feedback if he goes off in the wrong direction, giving you more control over tasks.
+- Re-imagined tool calling prompt resulting in ~40% fewer requests to accomplish tasks + better performance with other models.
+- Search and use any model with OpenRouter (try searching "free" for free model offerings).
 
 ## [1.9.7]